refactor: move manifest package out of internal/pkg
internal/pkg is a generic bucket the project is migrating away from.
Move the existing internal/pkg/manifest package (validate plus the
new sync/diff/merge/flatten/writeback/display logic) up to
internal/manifest and update its three importers.
