+++ internal/manifest/diff.go
