+++ internal/manifest/diff_test.go
