+++ internal/manifest/display.go
