+++ internal/manifest/flatten.go
