+++ internal/manifest/flatten_test.go
