+++ internal/manifest/merge.go
