+++ internal/manifest/merge_test.go
