+++ internal/manifest/sync.go
