+++ internal/manifest/validate.go
