+++ internal/manifest/validate_test.go
