+++ internal/manifest/writeback.go
